fix: prevent local changes from breaking compare (#129)
## Details

When external storage is used and there are local changes, the git CLI fails to execute `git stash pop` because `git stash` was never executed. This change uses a boolean to make sure we only pop if we stashed. This should fix lightning-components perf runs which are currently failing with `Error: No stash found`.

## Does this PR introduce a breaking change?

* [ ] Yes
* [x] No