fix: improve efficiency by not doing unneeded extracting/compressing
Unfortunately, because of the way Best was designed, the hub does
a lot of unnecessary¹ extracting and compressing of files, when it
already has the `tar` it needs to send to an agent.

Furthermore, because of the way the hub does the compression¹, with
every additional test, the `tar` gets bigger and bigger. So, the more
tests Best has to run, the slower things will be, and sometimes Best
will even crash.

This commit is a band-aid fix for the above until a more major
refactoring\rearchitecting of Best is done.

- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -

¹ When the hub receives a request to run some tests, it gets the `tar`
  for the first test from the client. Instead of passing that `tar`
  to an available agent, it extracts its content in a directory, and
  then compresses the content it just extracted into a `tar` and sends
  it to an agent.

  When it receives the `tar` for the next test, it extracts its content
  in the same directory as the first test, compresses everything in that
  directory including the files for the previous test, and then sends
  that to the agent.

  This goes on and on and on, the more tests Best has to run, the bigger
  the payload that it will have to compress and send to the agent that,
  in turn, will also have to extract.