fix: ensure files are fully written to disk before they are used
Best sometimes errors with `zlib: unexpected end of file` as the
`extractTar` function tries to extract the data from a corrupt `tar`
file.

What seems to be happening is that `socket.io-file` signals that
it completed the writing of the `tar` file on disk when it fact the
`writeStream` didn't actually finish. Hence, when `extractTar` starts
extracting the content of the `tar` file, it reads an incomplete file
and errors out.

This commit tries to fix the issue described above by making the
`socket.io-file` ensure the `writeStream` is closed before signaling
completion.